# coffee-machine-adapters/src/main/resources/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Coffee Machine API
  version: 1.0.0
  description: >
    REST API for brewing coffee, controlling power, refilling resources and querying
    status/recipes. This spec mirrors the current Spring MVC controllers.
servers:
  - url: http://localhost:8080
...
paths:
  /api/brew:
    post:
      tags: [Brew]
      operationId: startBrew
      summary: Start a brew job
      description: |
        Accepts a brew request for a specific recipe. Returns 202 even when the brew
        is rejected (payload indicates success).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartBrewRequest'
            examples:
              espresso:
                value: { recipeId: "11111111-2222-3333-4444-555555555555" }
      responses:
        '202':
          description: Brew request accepted by the controller
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartBrewResponse'
              examples:
                accepted:
                  value: { started: true, message: "Brewing Espresso for 20s" }
                busy:
                  value: { started: false, message: "Machine is already brewing" }
...
components:
  schemas:
    StartBrewRequest:
      type: object
      required: [recipeId]
      properties:
        recipeId:
          type: string
          format: uuid
          description: ID of the recipe to brew.
    StartBrewResponse:
      type: object
      required: [started, message]
      properties:
        started:
          type: boolean
          description: Whether the brew was accepted and started.
        message:
          type: string
          description: Human-readable message.
...